Modularize passport
