+++ server/authentication.js
@@ -1,0 +1,20 @@
+const passport = require('passport');
+const Strategy = require('passport-twitter').Strategy;
+
+passport.use(new Strategy({
+  consumerKey: process.env.TWITTER_API,
+  consumerSecret: process.env.TWITTER_SECRET,
+  callbackURL: 'http://127.0.0.1:8000/auth/twitter/callback',
+}, (token, tokenSecret, profile, done) => done(null, profile)));
+
+passport.serializeUser((user, done) => {
+  // console.log('serialize user in server: ', user)
+  done(null, user);
+});
+
+passport.deserializeUser((obj, done) => {
+  // console.log('deserialize obj in server: ', obj)
+  done(null, obj);
+});
+
+module.exports = passport;
